Moved various (currenty unused, but useful for debugging tests) FX methods out of TestUtil into TFXUtil.  Also removed some unused bits of TestUtil.
